create folder network utils + library
